[Security Solution][Endpoint] Split defend workflows cypress tests to minimize run times (#170232)
## Summary

Splits test suites into multiple files to minimize CI run time.

- [x] splits tests in
[response_console_actions.cy.ts](https://github.com/elastic/kibana/blob/main/x-pack/plugins/security_solution/public/management/cypress/e2e/response_actions/response_console_actions.cy.ts)
- [x] splits
[response_console_mocked_data.cy.ts](https://github.com/elastic/kibana/blob/main/x-pack/plugins/security_solution/public/management/cypress/e2e/response_actions/response_console_mocked_data.cy.ts)

### Checklist
- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios

(cherry picked from commit 2789dd78244df46781bcca53a141678fb8c166c3)

# Conflicts:
#	x-pack/plugins/security_solution/public/management/cypress/e2e/response_actions/response_console_actions.cy.ts
#	x-pack/plugins/security_solution/public/management/cypress/e2e/response_actions/response_console_mocked_data.cy.ts
